auth_service/requirements.txt
absl-py==1.0.0
aioredis==2.0.1
ale-py==0.8.0
anyio==3.5.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.0
astroid==2.3.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
atari-py==0.2.9
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
black==22.3.0
blinker==1.4
box2d-py==2.3.5
Brlapi==0.7.0
Brotli==1.0.9
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.0.0
cachy==0.3.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
cleo==0.8.1
click==8.1.2
clikit==0.6.2
cloudpickle==1.6.0
colorama==0.4.1
command-not-found==0.3
commonmark==0.9.1
contourpy==1.0.5
crashtest==0.3.1
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.28
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
evdev==1.3.0
fake-useragent==0.1.11
fastapi==0.75.2
fasteners==0.18
filelock==3.6.0
flake8==5.0.4
Flask==1.1.1
flatbuffers==2.0
fonttools==4.33.0
gast==0.5.3
glfw==2.5.5
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.44.0
gym==0.18.3
gym-cartpole-swingup==0.1.4
gym-notices==0.0.8
h11==0.13.0
h5py==3.6.0
hiredis==2.0.0
html5lib==1.1
httplib2==0.14.0
httptools==0.1.2
idna==2.8
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.3
importlib-resources==5.10.0
iniconfig==1.1.1
iso8601==1.0.2
isort==4.3.21
itsdangerous==1.1.0
jarowinkler==1.1.0
jeepney==0.8.0
Jinja2==2.10.3
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.2
libclang==14.0.1
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==4.0.2
m3u8==1.0.0
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.5.1
mccabe==0.7.0
moviepy==1.0.3
msgpack==1.0.3
mujoco==2.2.0
mujoco-py==2.1.2.14
mutagen==1.45.1
mypy-extensions==0.4.3
netifaces==0.10.4
networkx==2.4
netwulf==0.1.4
nic==0.0.1
nltk==3.7
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.2.0.34
opencv-python==4.5.5.64
opt-einsum==3.3.0
outcome==1.1.0
packaging==20.9
pafy==0.5.5
pandas==1.4.2
parse==1.15.0
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.1.0
pkginfo==1.8.2
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
poetry==1.1.13
poetry-core==1.0.8
poetry-version==0.1.5
pptree==3.1
proglog==0.1.10
protobuf==3.20.1
psutil==5.5.1
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.15.0
pycups==1.9.73
pydantic==1.9.0
pyee==7.0.1
pyflakes==2.5.0
pygame==2.1.0
pyglet==1.5.15
Pygments==2.12.0
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pylev==1.4.0
pyline==0.3.20
pylint==2.4.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
PyOpenGL==3.1.6
pyOpenSSL==19.0.0
pyparsing==3.0.8
pyppeteer==0.0.25
pyquery==1.4.1
pyRFC3339==1.1
PySimpleGUI==4.60.3
PySocks==1.7.1
pytest==7.1.3
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-louvain==0.13
python-magic==0.4.16
python-ulid==1.1.0
pytube==12.1.0
pytz==2022.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
rapidfuzz==2.1.2
redis==4.2.2
redis-om==0.0.24
regex==2022.3.15
reportlab==3.5.34
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rich==12.4.4
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==3.3.2
selenium==4.2.0
setproctitle==1.1.10
shellingham==1.4.0
simplejson==3.17.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotdl==3.9.6
spotipy==2.20.0
starlette==0.17.1
swig==4.1.0
systemd-python==234
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.10.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
twitch-dl==1.22.0
typed-ast==1.4.0
types-redis==4.1.22
types-six==1.16.15
typing-extensions==4.2.0
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.3.4
urllib3==1.26.9
usb-creator==0.3.7
uvicorn==0.17.6
uvloop==0.16.0
vboxapi==1.0
virtualenv==20.14.1
w3lib==1.21.0
wadllib==1.3.3
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.0
wrapt==1.14.0
wsproto==0.15.0
xkit==0.0.0
youtube-dl==2020.12.2
yt-dlp==2022.6.29
ytmusicapi==0.22.0
zipp==3.8.0
